+++ test/support/cache_helpers.ex
@@ -7,6 +7,7 @@ defmodule CacheHelpers do
   import ExUnit.Callbacks, only: [on_exit: 1]
 
   alias TdCache.HierarchyCache
+  alias TdCache.I18nCache
   alias TdCache.SystemCache
   alias TdCache.TaxonomyCache
   alias TdCache.TemplateCache
@@ -39,4 +40,11 @@ defmodule CacheHelpers do
     on_exit(fn -> HierarchyCache.delete(hierarchy_id) end)
     hierarchy
   end
+
+  def put_i18n_messages(lang, messages) when is_list(messages) do
+    Enum.each(messages, &I18nCache.put(lang, &1))
+    on_exit(fn -> I18nCache.delete(lang) end)
+  end
+
+  def put_i18n_message(lang, message), do: put_i18n_messages(lang, [message])
 end